(windows) removeDirectory able to force delete folders with content

+++ src/utilities/directorycreator.cpp
@@ -19,9 +19,11 @@ bool DirectoryCreator::createDirectory(const std::string& path)
     return mkdir(path.c_str(), 0755) != -1 ? true : false;
 }
 
-bool DirectoryCreator::removeDirectory(const std::string& path)
+bool DirectoryCreator::removeDirectory(const std::string& path, bool force)
 {
     return rmdir(path.c_str()) != -1 ? true : false;
+	
+	// implement force delete here too
 }
 
 #elif defined(__WIN32__)
@@ -38,9 +40,26 @@ bool DirectoryCreator::createDirectory(const std::string& path)
     return CreateDirectory(path.c_str(), NULL) != 0 ? true : false;
 }
 
-bool DirectoryCreator::removeDirectory(const std::string& path)
+bool DirectoryCreator::removeDirectory(const std::string& path, bool force)
 {
-    return RemoveDirectory(path.c_str());
+	if(!force)
+	{
+		return RemoveDirectory(path.c_str());
+	}
+	else
+	{
+		SHFILEOPSTRUCT file_op = {
+			NULL,
+			FO_DELETE,
+			(path + '\0').c_str(),
+			"",
+			FOF_NOCONFIRMATION | FOF_NOERRORUI | FOF_SILENT,
+			false,
+			0,
+			""};
+		
+		return !SHFileOperation(&file_op);
+	}
 }
 
 #else
